[FIX] website_quote:
This revision is related to 29923ca6a71c8df8775b48206e0d0d627d8fa915

The payment acquirers must be fetched if the quote
needs a payment.

Besides, as explained in the above revision, the state
`manual` no longer exists sales orders, and therefore
this must no longer be used to determine if a quote
needs a payment or not.

opw-660575