[FIX] base: url field for ir.module.module
The ir.module.module model has a 'url' field (so there is a 'url' field
in db) but the 'url' values in the __openerp__.py files were not used.

Now, the 'url' field is filled with the manifest 'url' value with a
fallback to the manifest 'live_test_url' value.