[FIX] account: multi-company payment term
Despite the fact there is a `company_id` field
on `account.payment.term`, there is no multi-company
record rule on this model. The payment terms
are therefore shared by default accross companies,
even if the company is filled on the payment term.

In addition:
 - The `company_id` field on `account.payment.term` is required
 - The field is invisible on the payment term form
 - The default value is the user company

Adding the record rule would imply to change:
 - The fact the `company_id` field is required,
   as you probably would like to share some payment
   terms accross companies, by leaving empty this field
   on the payment term form
 - The field should be visible in the form for users
   having the multi-company group, so the admins can
   create payment terms for other companies than the one
   they are currently in.

This cannot be done in a stable release, such as 9.0.
This is for master.

In the mean time, we keep the behavior as it was in 8.0:
We consider that the payment terms are shared among companies,
and the `company_id` field is simply ignored.

This implies that choosing a payment term from another company
in an invoice must not lead to multi-company access rights issues,
as it was the case before this revision: If a user other than
the admin chose a payment term from another company, the on change
on the payment term leaded to access rights issues, as it needed
the read access on the company of the payment term, to get
the currency decimal places.

From this revision, the currency is took from the invoice directly,
through the context, and if not passed in the context, the company
of the user is used instead of the company of the paymen term.
It makes more sense anyway to take the currency from the invoice
instead of the currency of the payment term company to retrieve
the currency decimal places.

opw-659818