[FIX] website_quote: correct company headers
When printing the online quote, using the "Print" button
in the online proposal in the front-end, the user
used to print the report is the SUPERUSER.

If `doc` or `o` is not set within the QWeb values,
the headers printed are the ones from the user company,
instead of the record company.

As this was the case in the `website_quote` report,
the company headers printed were not the sale order company
headers, but the SUPERUSER company headers.

opw-666306