[FIX] pos_mercury: load pos_mercury_config_id from account.journal
The introduction of cfa2eb564b1175c7c115e29fe2626cf44816e76b made it so
that only the type and sequence field of account.journal get loaded in
the POS. Before that patch all fields of account.journal would get
loaded. pos_mercury depended on this behavior, because it needs the
pos_mercury_config_id.

This loads the field explicitly.

Because of the change the display_name field on the journal is also
unavailable, so we instead use the journal_id field on
account.bank.statement. This is probably even better than what we did
before, because this is what we also use when printing.