[FIX] account, account_asset, purchase: extra field in auto fill
The Purchase module provides a functionality to add pre-filled invoice
lines to an invoice. The issue is that extensions of the accounting
modules might add fields on an invoice line, and one would like to see
these fields populated as well.

This is the case for account_asset, which adds an asset_category_id on
an invoice line. However, the Purchase module is not aware of the fields
added by account_asset, and therefore it cannot populate them.

This fix adds a method, _auto_fill_extra, which is intended to manage
these extra fields.

opw-667162