[FIX] calendar: To avoid duplicate contacts
Before the fix bd7f4cf00abcd38d0811ab295c06f5810f18c03f, the duplicate
records were erased thanks to the index used in the dictionary(id of the partner).
After this fix, to avoid this this problem, the label of each record is used as
key in the dictionary.