[FIX] mail: announcement bar URl displayed on print
On printing, the announcement bar URL was displayed
in the printing result.
e.g.
`https://services.openerp.com/openerp-enterprise/ab/register/...`
was displayed at the top of the print.

This is related to the revision 6b9e05ae21d54358402a71d901d2d11e59801b41.

This is due to this simple typo, the according CSS rule was
never applied because of it.